Changed type in structs to Float64

+++ src/Experiments/src/optimisers.jl
@@ -1,9 +1,9 @@
 @option struct OptConfig
     type::String = "Descent"
-    eta::Float32 = 0.01
+    eta::Float64 = 0.01
     decay_every::Int = 1
-    decay_step::Float32 = 1
-    decay_min::Float32 = 1e-6
+    decay_step::Float64 = 1
+    decay_min::Float64 = 1e-6
 end
 
 function Base.string(o::OptConfig)
@@ -12,5 +12,5 @@ function Base.string(o::OptConfig)
 end
 
 materialize(o::OptConfig) = materialize(Val(Symbol(o.type)), o)
-materialize(::Val{:Descent}, o::OptConfig) = Descent(o.eta)
-materialize(::Val{:ADAM}, o::OptConfig) = ADAM(o.eta)
+materialize(::Val{:Descent}, o::OptConfig) = Descent(Float32(o.eta))
+materialize(::Val{:ADAM}, o::OptConfig) = ADAM(Float32(o.eta))
